Acknowledgements added and PostGIS working
Added acknowledgements page and table.

PostGIS: added geom field, added the PostGIS extension, and populated the geom field with a pgAdmin sql. Probably better if I could figure how to do this in Rails.

Next need to confirm that the database is usable with QGIS and other apps.

Then add time factor which will require a join table with years.

Then add data.